Fixed a typo.

+++ project/Assets/Scripts/GoliathNetworking.cs
@@ -177,7 +177,7 @@ public class GoliathNetworking : Photon.MonoBehaviour {
 	}
 	
 	[RPC]
-	public void CreatePlayerBullet(float damage, Vector3 position, float speed, Vector3 direction){
+	public void CreatePlayerBullet(float damage, Vector3 position, int speed, Vector3 direction){
 		GameObject projectile = playerBulletManager.Retrieve(position);
 
 		Bullet bulletScript;
